+++ src/builtin/builtin.c
@@ -3,12 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   builtin.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: liurne <liurne@student.42.fr>              +#+  +:+       +#+        */
+/*   By: jcoquard <jcoquard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/25 15:27:21 by liurne            #+#    #+#             */
-/*   Updated: 2023/10/26 14:35:56 by liurne           ###   ########.fr       */
+/*   Updated: 2023/10/30 15:59:09 by jcoquard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
+int	is_builtin(t_cmd cmd)
+{
+	
+}
