+++ src/process.c
@@ -6,7 +6,7 @@
 /*   By: jcoquard <jcoquard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/27 17:05:37 by jcoquard          #+#    #+#             */
-/*   Updated: 2023/10/27 19:32:34 by jcoquard         ###   ########.fr       */
+/*   Updated: 2023/10/30 15:09:04 by jcoquard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,11 +17,12 @@ int	set_rval(int val, char *error)
 	if (error)
 		ft_dprintf(2, "%s", error);
 	g_rvalue = val;
+	return (val);
 }
 
 int	process(t_data *shell)
 {
-	int	i;
+	unsigned int	i;
 
 	i = -1;
 	if (striswspace(shell->prompt.line))
@@ -33,5 +34,9 @@ int	process(t_data *shell)
 	while (++i < shell->prompt.nb_cmds)
 		shell->prompt.cmds[i].cmd = expand(shell, &(shell->prompt.cmds[i]),
 				shell->prompt.cmds[i].cmd);
+	i = -1;
+	while (++i < shell->prompt.nb_cmds)
+		exec(shell, &(shell->prompt.cmds[i]));
+	free_cmds(shell);
 	return (0);
 }
