+++ src/utils/strcpy_neg.c
